Extracting method to ease future switching logic
In discussion with @no_reply, in addressing #4195, we want to explore
the following path:

1) As early as possible in the request cycle, create a
   Hyrax::UploadedFile for both a CarrierWave file or a file selected
   for upload by BrowseEverything.
2) Building on step 1, we would then hope to reduce branching logic in the
   request cycle, and preserve existing class interfaces.

By making this non-breaking change, I'm orienting the code to deliver on
this aspirational goal.

Next steps would be:

1) Update the Hyrax::WorksControllerBehavior to create
   Hyrax::UploadedFile records for the files selected from
   BrowseEverything (note CarrierWave handles the "inline" uploads).
2) Adjust the extracted `Hyrax::UploadedFile#perform_ingest_later`
   method to handle files that were created from BrowseEverything.

The plan is to add a field to the UploadedFile (perhaps
`url_of_file_to_ingest`) that the Hyrax::WorksControllerBehavior would
populate based on user input parameters.

With that in place, we could look at backporting the solution into the
ActorStack.